Refactors Auth0 rules
Each rule now does one thing only, and is more clearly named
